Practical Machine Learning Assignment: Prediction Assignment Writeup
One thing that people regularly do is quantify how much of a particular activity they do, but they rarely quantify how well they do it. In this project, the data from accelerometers on the belt, forearm, arm, and dumbell of 6 participants is used to predict the manner in which the participants did the exercise.